# backend/src/swagger.json
{
	"openapi": "3.0.0",
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"_36_Enums.UserRole": {
				"type": "string",
				"enum": [
					"SUPER_ADMIN",
					"OWNER",
					"MANAGER",
					"LIBRARIAN",
					"USER",
					"GUEST"
				]
			},
			"_36_Enums.MembershipType": {
				"type": "string",
				"enum": [
					"STANDARD",
					"PREMIUM",
					"VIP"
				]
			},
			"UserRole": {
				"$ref": "#/components/schemas/_36_Enums.UserRole"
			},
			"CreateUserDto": {
				"properties": {
					"phone": {
						"type": "number",
						"format": "double"
					},
					"password": {
						"type": "string"
					},
					"fullName": {
						"type": "string"
					},
					"avatar": {
						"type": "string"
					},
					"role": {
						"$ref": "#/components/schemas/UserRole"
					}
				},
				"required": [
					"phone",
					"password",
					"fullName"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UpdateUserDto": {
				"properties": {
					"phone": {
						"type": "string"
					},
					"password": {
						"type": "string"
					},
					"fullName": {
						"type": "string"
					},
					"avatar": {
						"type": "string"
					},
					"role": {
						"$ref": "#/components/schemas/UserRole"
					},
					"isActive": {
						"type": "boolean"
					},
					"isVerified": {
						"type": "boolean"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"LoginDto": {
				"description": "Login credentials DTO",
				"properties": {
					"phone": {
						"type": "string",
						"description": "User's phone number",
						"example": "+998901234567"
					},
					"password": {
						"type": "string",
						"description": "User's password",
						"example": "password123"
					}
				},
				"required": [
					"phone",
					"password"
				],
				"type": "object",
				"additionalProperties": false
			},
			"RegisterDto": {
				"description": "Registration details DTO",
				"properties": {
					"phone": {
						"type": "string",
						"description": "User's phone number",
						"example": "+998901234567"
					},
					"fullName": {
						"type": "string",
						"description": "User's Full name",
						"example": "John Doe"
					},
					"password": {
						"type": "string",
						"description": "User's password",
						"example": "password123"
					}
				},
				"required": [
					"phone",
					"fullName",
					"password"
				],
				"type": "object",
				"additionalProperties": false
			},
			"SendCodeDto": {
				"description": "Send verification code DTO",
				"properties": {
					"phone": {
						"type": "string",
						"description": "User's phone number",
						"example": "+998901234567"
					}
				},
				"required": [
					"phone"
				],
				"type": "object",
				"additionalProperties": false
			},
			"VerifyCodeDto": {
				"description": "Verify code DTO",
				"properties": {
					"phone": {
						"type": "string",
						"description": "User's phone number",
						"example": "+998901234567"
					},
					"code": {
						"type": "string",
						"description": "6-digit verification code",
						"example": "123456"
					}
				},
				"required": [
					"phone",
					"code"
				],
				"type": "object",
				"additionalProperties": false
			},
			"LogOutDto": {
				"description": "Log Out DTO",
				"properties": {
					"refreshToken": {
						"type": "string",
						"description": "User's refresh token",
						"example": "refresh_token"
					}
				},
				"required": [
					"refreshToken"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "edujavon-crm",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {
			"name": "ozim"
		}
	},
	"paths": {
		"/users": {
			"get": {
				"operationId": "GetUsers",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"properties": {
											"updatedAt": {
												"type": "string",
												"format": "date-time"
											},
											"createdAt": {
												"type": "string",
												"format": "date-time"
											},
											"deletedAt": {
												"type": "string",
												"format": "date-time"
											},
											"isActive": {
												"type": "boolean"
											},
											"isVerified": {
												"type": "boolean"
											},
											"telegramChatId": {
												"type": "string"
											},
											"role": {
												"$ref": "#/components/schemas/_36_Enums.UserRole"
											},
											"avatar": {
												"type": "string"
											},
											"fullName": {
												"type": "string"
											},
											"phone": {
												"type": "string"
											},
											"id": {
												"type": "string"
											}
										},
										"required": [
											"updatedAt",
											"createdAt",
											"deletedAt",
											"isActive",
											"isVerified",
											"telegramChatId",
											"role",
											"avatar",
											"fullName",
											"phone",
											"id"
										],
										"type": "object"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Users"
				],
				"security": [],
				"parameters": []
			},
			"post": {
				"operationId": "CreateUser",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"properties": {
												"data": {},
												"message": {
													"type": "string"
												}
											},
											"required": [
												"message"
											],
											"type": "object"
										},
										{
											"properties": {
												"data": {
													"properties": {
														"updatedAt": {
															"type": "string",
															"format": "date-time"
														},
														"createdAt": {
															"type": "string",
															"format": "date-time"
														},
														"deletedAt": {
															"type": "string",
															"format": "date-time"
														},
														"totalPoints": {
															"type": "number",
															"format": "double"
														},
														"currentLevel": {
															"type": "number",
															"format": "double"
														},
														"booksReadCount": {
															"type": "number",
															"format": "double"
														},
														"lastLoginAt": {
															"type": "string",
															"format": "date-time"
														},
														"joinDate": {
															"type": "string",
															"format": "date-time"
														},
														"membershipExpiry": {
															"type": "string",
															"format": "date-time"
														},
														"membershipType": {
															"$ref": "#/components/schemas/_36_Enums.MembershipType"
														},
														"blockedUntil": {
															"type": "string",
															"format": "date-time"
														},
														"blockedReason": {
															"type": "string"
														},
														"isBlocked": {
															"type": "boolean"
														},
														"isActive": {
															"type": "boolean"
														},
														"isVerified": {
															"type": "boolean"
														},
														"city": {
															"type": "string"
														},
														"address": {
															"type": "string"
														},
														"dateOfBirth": {
															"type": "string",
															"format": "date-time"
														},
														"telegramVerified": {
															"type": "boolean"
														},
														"telegramLinkedAt": {
															"type": "string",
															"format": "date-time"
														},
														"telegramUsername": {
															"type": "string"
														},
														"telegramChatId": {
															"type": "string"
														},
														"telegramId": {
															"type": "string"
														},
														"role": {
															"$ref": "#/components/schemas/_36_Enums.UserRole"
														},
														"avatar": {
															"type": "string"
														},
														"fullName": {
															"type": "string"
														},
														"password": {
															"type": "string"
														},
														"phone": {
															"type": "string"
														},
														"email": {
															"type": "string"
														},
														"id": {
															"type": "string"
														}
													},
													"required": [
														"updatedAt",
														"createdAt",
														"deletedAt",
														"totalPoints",
														"currentLevel",
														"booksReadCount",
														"lastLoginAt",
														"joinDate",
														"membershipExpiry",
														"membershipType",
														"blockedUntil",
														"blockedReason",
														"isBlocked",
														"isActive",
														"isVerified",
														"city",
														"address",
														"dateOfBirth",
														"telegramVerified",
														"telegramLinkedAt",
														"telegramUsername",
														"telegramChatId",
														"telegramId",
														"role",
														"avatar",
														"fullName",
														"password",
														"phone",
														"email",
														"id"
													],
													"type": "object"
												},
												"message": {
													"type": "string"
												}
											},
											"required": [
												"data",
												"message"
											],
											"type": "object"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Users"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateUserDto"
							}
						}
					}
				}
			}
		},
		"/users/{id}": {
			"get": {
				"operationId": "GetUserById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"properties": {
												"data": {},
												"message": {
													"type": "string"
												}
											},
											"required": [
												"message"
											],
											"type": "object"
										},
										{
											"properties": {
												"message": {},
												"data": {
													"properties": {
														"updatedAt": {
															"type": "string",
															"format": "date-time"
														},
														"createdAt": {
															"type": "string",
															"format": "date-time"
														},
														"deletedAt": {
															"type": "string",
															"format": "date-time"
														},
														"totalPoints": {
															"type": "number",
															"format": "double"
														},
														"currentLevel": {
															"type": "number",
															"format": "double"
														},
														"booksReadCount": {
															"type": "number",
															"format": "double"
														},
														"lastLoginAt": {
															"type": "string",
															"format": "date-time"
														},
														"joinDate": {
															"type": "string",
															"format": "date-time"
														},
														"membershipExpiry": {
															"type": "string",
															"format": "date-time"
														},
														"membershipType": {
															"$ref": "#/components/schemas/_36_Enums.MembershipType"
														},
														"blockedUntil": {
															"type": "string",
															"format": "date-time"
														},
														"blockedReason": {
															"type": "string"
														},
														"isBlocked": {
															"type": "boolean"
														},
														"isActive": {
															"type": "boolean"
														},
														"isVerified": {
															"type": "boolean"
														},
														"city": {
															"type": "string"
														},
														"address": {
															"type": "string"
														},
														"dateOfBirth": {
															"type": "string",
															"format": "date-time"
														},
														"telegramVerified": {
															"type": "boolean"
														},
														"telegramLinkedAt": {
															"type": "string",
															"format": "date-time"
														},
														"telegramUsername": {
															"type": "string"
														},
														"telegramChatId": {
															"type": "string"
														},
														"telegramId": {
															"type": "string"
														},
														"role": {
															"$ref": "#/components/schemas/_36_Enums.UserRole"
														},
														"avatar": {
															"type": "string"
														},
														"fullName": {
															"type": "string"
														},
														"password": {
															"type": "string"
														},
														"phone": {
															"type": "string"
														},
														"email": {
															"type": "string"
														},
														"id": {
															"type": "string"
														}
													},
													"required": [
														"updatedAt",
														"createdAt",
														"deletedAt",
														"totalPoints",
														"currentLevel",
														"booksReadCount",
														"lastLoginAt",
														"joinDate",
														"membershipExpiry",
														"membershipType",
														"blockedUntil",
														"blockedReason",
														"isBlocked",
														"isActive",
														"isVerified",
														"city",
														"address",
														"dateOfBirth",
														"telegramVerified",
														"telegramLinkedAt",
														"telegramUsername",
														"telegramChatId",
														"telegramId",
														"role",
														"avatar",
														"fullName",
														"password",
														"phone",
														"email",
														"id"
													],
													"type": "object"
												}
											},
											"required": [
												"data"
											],
											"type": "object"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Users"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"put": {
				"operationId": "UpdateUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"properties": {
												"data": {},
												"message": {
													"type": "string"
												}
											},
											"required": [
												"message"
											],
											"type": "object"
										},
										{
											"properties": {
												"data": {
													"properties": {
														"updatedAt": {
															"type": "string",
															"format": "date-time"
														},
														"createdAt": {
															"type": "string",
															"format": "date-time"
														},
														"deletedAt": {
															"type": "string",
															"format": "date-time"
														},
														"isActive": {
															"type": "boolean"
														},
														"isVerified": {
															"type": "boolean"
														},
														"telegramChatId": {
															"type": "string"
														},
														"role": {
															"$ref": "#/components/schemas/_36_Enums.UserRole"
														},
														"avatar": {
															"type": "string"
														},
														"fullName": {
															"type": "string"
														},
														"phone": {
															"type": "string"
														},
														"id": {
															"type": "string"
														}
													},
													"required": [
														"updatedAt",
														"createdAt",
														"deletedAt",
														"isActive",
														"isVerified",
														"telegramChatId",
														"role",
														"avatar",
														"fullName",
														"phone",
														"id"
													],
													"type": "object"
												},
												"message": {
													"type": "string"
												}
											},
											"required": [
												"data",
												"message"
											],
											"type": "object"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Users"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateUserDto"
							}
						}
					}
				}
			}
		},
		"/users/hard/{id}": {
			"delete": {
				"operationId": "HardDelete",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"properties": {
												"data": {},
												"message": {
													"type": "string"
												}
											},
											"required": [
												"message"
											],
											"type": "object"
										},
										{
											"properties": {
												"data": {
													"properties": {
														"updatedAt": {
															"type": "string",
															"format": "date-time"
														},
														"createdAt": {
															"type": "string",
															"format": "date-time"
														},
														"deletedAt": {
															"type": "string",
															"format": "date-time"
														},
														"totalPoints": {
															"type": "number",
															"format": "double"
														},
														"currentLevel": {
															"type": "number",
															"format": "double"
														},
														"booksReadCount": {
															"type": "number",
															"format": "double"
														},
														"lastLoginAt": {
															"type": "string",
															"format": "date-time"
														},
														"joinDate": {
															"type": "string",
															"format": "date-time"
														},
														"membershipExpiry": {
															"type": "string",
															"format": "date-time"
														},
														"membershipType": {
															"$ref": "#/components/schemas/_36_Enums.MembershipType"
														},
														"blockedUntil": {
															"type": "string",
															"format": "date-time"
														},
														"blockedReason": {
															"type": "string"
														},
														"isBlocked": {
															"type": "boolean"
														},
														"isActive": {
															"type": "boolean"
														},
														"isVerified": {
															"type": "boolean"
														},
														"city": {
															"type": "string"
														},
														"address": {
															"type": "string"
														},
														"dateOfBirth": {
															"type": "string",
															"format": "date-time"
														},
														"telegramVerified": {
															"type": "boolean"
														},
														"telegramLinkedAt": {
															"type": "string",
															"format": "date-time"
														},
														"telegramUsername": {
															"type": "string"
														},
														"telegramChatId": {
															"type": "string"
														},
														"telegramId": {
															"type": "string"
														},
														"role": {
															"$ref": "#/components/schemas/_36_Enums.UserRole"
														},
														"avatar": {
															"type": "string"
														},
														"fullName": {
															"type": "string"
														},
														"password": {
															"type": "string"
														},
														"phone": {
															"type": "string"
														},
														"email": {
															"type": "string"
														},
														"id": {
															"type": "string"
														}
													},
													"required": [
														"updatedAt",
														"createdAt",
														"deletedAt",
														"totalPoints",
														"currentLevel",
														"booksReadCount",
														"lastLoginAt",
														"joinDate",
														"membershipExpiry",
														"membershipType",
														"blockedUntil",
														"blockedReason",
														"isBlocked",
														"isActive",
														"isVerified",
														"city",
														"address",
														"dateOfBirth",
														"telegramVerified",
														"telegramLinkedAt",
														"telegramUsername",
														"telegramChatId",
														"telegramId",
														"role",
														"avatar",
														"fullName",
														"password",
														"phone",
														"email",
														"id"
													],
													"type": "object"
												},
												"message": {
													"type": "string"
												}
											},
											"required": [
												"data",
												"message"
											],
											"type": "object"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Users"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/users/soft/{id}": {
			"delete": {
				"operationId": "SoftDelete",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"properties": {
												"data": {},
												"message": {
													"type": "string"
												}
											},
											"required": [
												"message"
											],
											"type": "object"
										},
										{
											"properties": {
												"data": {
													"properties": {
														"updatedAt": {
															"type": "string",
															"format": "date-time"
														},
														"createdAt": {
															"type": "string",
															"format": "date-time"
														},
														"deletedAt": {
															"type": "string",
															"format": "date-time"
														},
														"totalPoints": {
															"type": "number",
															"format": "double"
														},
														"currentLevel": {
															"type": "number",
															"format": "double"
														},
														"booksReadCount": {
															"type": "number",
															"format": "double"
														},
														"lastLoginAt": {
															"type": "string",
															"format": "date-time"
														},
														"joinDate": {
															"type": "string",
															"format": "date-time"
														},
														"membershipExpiry": {
															"type": "string",
															"format": "date-time"
														},
														"membershipType": {
															"$ref": "#/components/schemas/_36_Enums.MembershipType"
														},
														"blockedUntil": {
															"type": "string",
															"format": "date-time"
														},
														"blockedReason": {
															"type": "string"
														},
														"isBlocked": {
															"type": "boolean"
														},
														"isActive": {
															"type": "boolean"
														},
														"isVerified": {
															"type": "boolean"
														},
														"city": {
															"type": "string"
														},
														"address": {
															"type": "string"
														},
														"dateOfBirth": {
															"type": "string",
															"format": "date-time"
														},
														"telegramVerified": {
															"type": "boolean"
														},
														"telegramLinkedAt": {
															"type": "string",
															"format": "date-time"
														},
														"telegramUsername": {
															"type": "string"
														},
														"telegramChatId": {
															"type": "string"
														},
														"telegramId": {
															"type": "string"
														},
														"role": {
															"$ref": "#/components/schemas/_36_Enums.UserRole"
														},
														"avatar": {
															"type": "string"
														},
														"fullName": {
															"type": "string"
														},
														"password": {
															"type": "string"
														},
														"phone": {
															"type": "string"
														},
														"email": {
															"type": "string"
														},
														"id": {
															"type": "string"
														}
													},
													"required": [
														"updatedAt",
														"createdAt",
														"deletedAt",
														"totalPoints",
														"currentLevel",
														"booksReadCount",
														"lastLoginAt",
														"joinDate",
														"membershipExpiry",
														"membershipType",
														"blockedUntil",
														"blockedReason",
														"isBlocked",
														"isActive",
														"isVerified",
														"city",
														"address",
														"dateOfBirth",
														"telegramVerified",
														"telegramLinkedAt",
														"telegramUsername",
														"telegramChatId",
														"telegramId",
														"role",
														"avatar",
														"fullName",
														"password",
														"phone",
														"email",
														"id"
													],
													"type": "object"
												},
												"message": {
													"type": "string"
												}
											},
											"required": [
												"data",
												"message"
											],
											"type": "object"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Users"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/auth/login": {
			"post": {
				"operationId": "Login",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"description": "Foydalanuvchi tizimga kirishi",
				"tags": [
					"Auth"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"description": "Login ma'lumotlari",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/LoginDto",
								"description": "Login ma'lumotlari"
							}
						}
					}
				}
			}
		},
		"/auth/register": {
			"post": {
				"operationId": "Register",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"description": "Yangi foydalanuvchini ro'yxatdan o'tkazish.\r\nAvval telefon raqamini Telegram bot orqali tasdiqlash talab qilinadi.\r\nAgar foydalanuvchi botda raqamini ulashgan bo'lsa, chatId avtomatik bog'lanadi.",
				"tags": [
					"Auth"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"description": "Ro'yxatdan o'tish ma'lumotlari",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/RegisterDto",
								"description": "Ro'yxatdan o'tish ma'lumotlari"
							}
						}
					}
				}
			}
		},
		"/auth/send-code": {
			"post": {
				"operationId": "SendCode",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"description": "Telegram bot orqali tasdiqlash kodini yuborish.\r\nFoydalanuvchi avval botga kirib /start bosgan bo'lishi kerak.",
				"tags": [
					"Auth"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"description": "Telefon raqami",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/SendCodeDto",
								"description": "Telefon raqami"
							}
						}
					}
				}
			}
		},
		"/auth/verify-code": {
			"post": {
				"operationId": "VerifyCode",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"description": "Telegram bot orqali yuborilgan kodni tekshirish.\r\nKod to'g'ri bo'lsa, foydalanuvchiga ro'yxatdan o'tish uchun ruxsat beriladi.",
				"tags": [
					"Auth"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"description": "Telefon raqami va tasdiqlash kodi",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/VerifyCodeDto",
								"description": "Telefon raqami va tasdiqlash kodi"
							}
						}
					}
				}
			}
		},
		"/auth/admin-test": {
			"get": {
				"operationId": "AdminTest",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"description": "Admin paneli uchun test marshruti",
				"tags": [
					"Auth"
				],
				"security": [
					{
						"jwt": [
							"admin"
						]
					}
				],
				"parameters": []
			}
		},
		"/auth/logout": {
			"post": {
				"operationId": "LogOut",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"description": "Foydalanuvchi tizimdan chiqishi",
				"tags": [
					"Auth"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"description": "Faqatgina refresh token",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/LogOutDto",
								"description": "Faqatgina refresh token"
							}
						}
					}
				}
			}
		}
	},
	"servers": [
		{
			"url": "/api/v1"
		}
	]
}